Modified routings, moved controllers below /admin and created a new controller
