+++ app/assets/javascripts/admin/menus.coffee
