+++ app/assets/stylesheets/admin/menus.scss
@@ -1,3 +1,3 @@
-// Place all the styles related to the menus controller here.
+// Place all the styles related to the admin/menus controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: http://sass-lang.com/
