+++ app/controllers/admin/menus_controller.rb
@@ -1,0 +1,4 @@
+class Admin::MenusController < ApplicationController
+  def index
+  end
+end
