+++ app/controllers/admin/restaurants_controller.rb
@@ -1,0 +1,4 @@
+class Admin::RestaurantsController < ApplicationController
+  def index
+  end
+end
