+++ app/controllers/admin/tags_controller.rb
@@ -1,4 +1,4 @@
-class TagsController < ApplicationController
+class Admin::TagsController < ApplicationController
   def new
     @tag = Tag.new
   end
