--- app/controllers/menus_controller.rb
@@ -1,4 +1,0 @@
-class MenusController < ApplicationController
-  def index
-  end
-end
