--- app/controllers/restaurants_controller.rb
@@ -1,4 +1,0 @@
-class RestaurantsController < ApplicationController
-  def index
-  end
-end
