+++ app/helpers/admin/menus_helper.rb
@@ -1,0 +1,2 @@
+module Admin::MenusHelper
+end
