--- app/helpers/menus_helper.rb
@@ -1,2 +1,0 @@
-module MenusHelper
-end
