+++ app/views/admin/categories/index.html.erb
@@ -1,0 +1,1 @@
+<h1>All categories</h1>
