+++ app/views/admin/categories/new.html.erb
