+++ app/views/admin/categories/show.html.erb
