+++ app/views/admin/restaurants/index.html.erb
