+++ app/views/categories/index.html.erb
@@ -1,1 +1,2 @@
-<h1>All categories</h1>
+<h1>Categories#index</h1>
+<p>Find me in app/views/categories/index.html.erb</p>
