+++ spec/controllers/admin/menus_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Admin::MenusController, type: :controller do
+
+end
