+++ spec/controllers/categories_controller_spec.rb
@@ -2,13 +2,11 @@ require 'rails_helper'
 
 RSpec.describe CategoriesController, type: :controller do
 
-  describe "GET #new" do
+  describe "GET #index" do
     it "returns http success" do
-      get :new
+      get :index
       expect(response).to have_http_status(:success)
     end
   end
 
 end
-
-# TODO: categories_controller_spec
